something something
I don't know if I did this right but I coded in the movement and the title screen in GMS2, as well as the direction sprite changes when changing direction.

Not sure if this is how github should present it, I just literally uploaded the entire GMS2 draft in the repository folder. Do tell if there's something wrong

Co-Authored-By: TehAwesomestKitteh <7497912+TehAwesomestKitteh@users.noreply.github.com>